--- a/examples/nodeports/README.md
+++ b/examples/nodeports/README.md
@@ -1,0 +1,3 @@
+# Wasm NodePort plugin
+
+This is [NodePort plugin](https://github.com/kubernetes/kubernetes/tree/master/pkg/scheduler/framework/plugins/nodeports) implemented with the wasm extension.