… examples/nodeports/go.sum
… examples/nodeports/main.wasm