<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind feature

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:
Add NodePorts plugin, ported from the in-tree plugin. 

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->

Part of #101 

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### What are the benchmark results of this change?
<!--
If documentation-only or trivial, just write "N/A". Otherwise, run the following:

1. `make bench-example > before.txt` on the commit prior to your changes
2. `make bench-example > after.txt` on the commit of your changes
3. Paste the output of `benchstat before.txt after.txt` as a code block.

If you haven't yet installed benchstat, it looks like this.
```
$ go install golang.org/x/perf/cmd/benchstat@latest
```

-->
```benchstat
                                      │  before.txt  │             after.txt              │
                                      │    sec/op    │    sec/op     vs base              │
Example_NodeNumber/Simple/New-8         305.8m ± 10%   326.9m ± 21%       ~ (p=0.093 n=6)
Example_NodeNumber/Simple/Run-8         104.5µ ± 10%   111.6µ ±  9%       ~ (p=0.132 n=6)
Example_NodeNumber/Simple_Log/New-8     310.7m ±  7%   306.3m ±  6%       ~ (p=0.485 n=6)
Example_NodeNumber/Simple_Log/Run-8     119.0µ ± 14%   119.2µ ± 11%       ~ (p=0.937 n=6)
Example_NodeNumber/Advanced/New-8       658.6m ± 22%   713.5m ± 10%       ~ (p=0.394 n=6)
Example_NodeNumber/Advanced/Run-8       38.26µ ±  9%   38.48µ ±  0%       ~ (p=0.132 n=6)
Example_NodeNumber/Advanced_Log/New-8   676.7m ±  2%   657.2m ±  1%  -2.87% (p=0.026 n=6)
Example_NodeNumber/Advanced_Log/Run-8   45.47µ ±  6%   45.25µ ±  3%       ~ (p=0.394 n=6)
geomean                                 5.562m         5.683m        +2.17%

```